sap/requirements.txt
Babel==2.9.1
backcall==0.2.0
beautifulsoup4==4.9.3
bleach==3.3.1
bleach-allowlist==1.0.3
boto3==1.17.112
botocore==1.20.112
braintree==4.8.0
cachetools==4.2.4
cairocffi==1.2.0
CairoSVG==2.5.2
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
click==7.1.2
croniter==1.0.15
cryptography==3.4.8
cssselect==1.1.0
cssselect2==0.4.1
cssutils==2.4.0
# Editable Git install with no remote (cusweb==0.0.1)
-e /home/ahmed/Documents/frappe-bench/apps/cusweb
dataclasses-json==0.5.6
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
docopt==0.6.2
dropbox==11.7.0
-e git+https://nado_rml@bitbucket.org/EngPeterMaged/e_invoicing.git@99628e32bc6ec05cf7f2b66a73bedaf59277b1cf#egg=e_invoicing
# Editable Git install with no remote (ecomm==0.0.1)
-e /home/ahmed/Documents/frappe-bench/apps/ecomm
email-reply-parser==0.5.12
-e git+https://github.com/frappe/erpnext@4d8d4acec4ff486505f1e99c53e3c7ebdd9944a7#egg=erpnext
et-xmlfile==1.1.0
-e git+https://github.com/frappe/frappe.git@8cb00ffc03aac64fe82f94598273a9612ea376e7#egg=frappe
git-url-parse==1.2.2
gitdb==4.0.9
GitPython==3.1.27
gocardless-pro==1.22.0
google-api-core==1.31.5
google-api-python-client==2.2.0
google-auth==1.29.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.54.0
googlemaps==4.4.7
gunicorn==20.1.0
hiredis==2.0.0
html2text==2020.1.16
html5lib==1.1
httplib2==0.20.4
idna==2.10
install==1.3.5
ipython==7.31.1
isodate==0.6.1
jedi==0.18.1
Jinja2==3.0.3
jmespath==0.10.0
jsonobject==0.9.10
ldap3==2.9.1
lxml==4.8.0
markdown2==2.4.2
MarkupSafe==2.1.0
marshmallow==3.14.1
marshmallow-enum==1.5.1
matplotlib-inline==0.1.3
maxminddb==2.2.0
maxminddb-geolite2==2018.703
mypy-extensions==0.4.3
num2words==0.5.10
numpy==1.22.2
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.3
pandas==1.1.5
parso==0.8.3
passlib==1.7.4
paytmchecksum==1.7.0
pbr==5.8.1
pdfkit==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
plaid-python==7.2.1
ply==3.11
premailer==3.8.0
prompt-toolkit==3.0.28
protobuf==3.19.4
psutil==5.8.0
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==20.7.3
pycparser==2.21
pycryptodome==3.10.4
PyGithub==1.55
Pygments==2.11.2
PyJWT==2.0.1
PyMySQL==1.0.2
PyNaCl==1.5.0
pyOpenSSL==20.0.1
pyotp==2.6.0
pyparsing==3.0.7
PyPDF2==1.26.0
pyphen==0.12.0
PyPika==0.48.8
pypng==0.0.21
PyQRCode==1.2.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.8.2
python-stdnum==1.17
python-youtube==0.8.1
pytz==2021.1
PyYAML==5.4.1
rauth==0.7.3
razorpay==1.2.0
redis==3.5.3
redisearch==2.0.0
requests==2.25.1
requests-oauthlib==1.3.1
RestrictedPython==5.2
rmtest==0.7.0
rq==1.8.1
rsa==4.8
s3transfer==0.4.2
-e git+https://github.com/ahmed751995/qr-generator.git@37d8996d2d75db747a47e4922d6bdd9b6f435d7e#egg=sap
# Editable Git install with no remote (scaler==0.0.1)
-e /home/ahmed/Documents/frappe-bench/apps/scaler
schedule==1.1.0
semantic-version==2.8.5
six==1.16.0
smmap==5.0.0
soupsieve==2.3.1
sqlparse==0.4.2
stone==3.3.1
stripe==2.56.0
taxjar==1.9.2
tenacity==8.0.1
terminaltables==3.1.10
tinycss2==1.1.1
traitlets==5.1.1
tweepy==3.10.0
typing-inspect==0.7.1
typing_extensions==4.1.1
Unidecode==1.2.0
uritemplate==3.0.1
urllib3==1.26.8
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
Werkzeug==2.0.3
Whoosh==2.7.4
wrapt==1.12.1
xlrd==2.0.1
zxcvbn-python==4.4.24
